Defined most PARAMs Made so the PDOException is properly catched

+++ base.sql
@@ -3,49 +3,49 @@ use requerimentos;
 
 
 create table administradores(
-    siape numeric(7) primary key,
+    siape int(7) primary key,
 
     nome char(255) not null,
     email char(255) not null,
-    curso numeric(1) not null,
+    curso int(1) not null,
 
     psswd binary(32) not null
 );
 
 
 create table discente(
-    matricula numeric(12) primary key,
+    matricula int(12) primary key,
 
     nome char(255) not null,
     email char(255) not null,
-    fone numeric(11) not null,
-    curso numeric(1) not null,
-    turma numeric(1) not null,
+    fone int(11) not null,
+    curso int(1) not null,
+    turma int(1) not null,
 
     psswd binary(32) not null
 );
 
 
 create table requerimento(
-    id numeric primary key auto,
+    id int primary key auto,
 
-    discente_id numeric(12) foreign key discente(matricula),
+    discente_id int(12) foreign key discente(matricula),
 
-    objeto numeric(1) not null,
+    objeto int(1) not null,
     inicio date not null,
     termino date not null,
     registro datetime default(getdate()) not null,
 
     anexo varbinary(max) not null, 
     obs char(255),
-    situacao numeric(1) not null
+    situacao int(1) not null
 );
 
 
 create table falta(
-    requerimento_id numeric foreign key requerimento(id),
-    docente_id numeric foreign key docente(id),
-    faltas numeric(1)
+    requerimento_id int foreign key requerimento(id),
+    docente_id int foreign key docente(id),
+    faltas int(1)
 );
